Hit with bug [JavaFX window stops being resizable after a dialog is shown]
https://bugs.openjdk.java.net/browse/JDK-8140491

As per the document, this only happens in Ubuntu but not on Windows or Mac.
However, the same bug appears in Fedora (I'm using Fedora). Targeted fix
by the java masters is on Java 10???

In this case, progress dialog upon opening up the database browser
window activates the bug.

For now, the progress dialog is disabled until we can move to other kind
of strategy. Maybe a progress indicator at the window itself.